backend/requirements.txt
django==5.1.1
djangorestframework==3.15.2
mysqlclient==2.2.4
python-decouple==3.8
gunicorn==23.0.0
openpyxl==3.1.5
django-cors-headers==4.4.0